requirements.txt
absl-py==2.2.2
accelerate==1.1.1
aiofiles==23.2.1
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
consisid_eva_clip==1.0.2
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
decord==0.6.0
deepspeed==0.15.2
defusedxml==0.7.1
diffusers==0.34.0.dev0
docker-pycreds==0.4.0
easydict==1.13
einops==0.8.1
executing==2.2.0
facexlib==0.3.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
ftfy==6.3.1
func_timeout==4.3.5
gitdb==4.0.12
GitPython==3.1.44
gradio==5.6.0
gradio_client==1.4.3
grpcio==1.71.0
h11==0.16.0
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.27.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
insightface==0.7.3
iopath==0.1.10
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.43.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
ninja==1.11.1.4
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.5.0
omegaconf==2.3.0
onnx==1.17.0
onnxruntime-gpu==1.19.2
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
orjson==3.10.18
packaging==25.0
pandas==2.2.3
parso==0.8.4
peft==0.12.0
pexpect==4.9.0
pillow==10.4.0
pip==25.1
platformdirs==4.3.7
portalocker==3.1.1
prettytable==3.16.0
proglog==0.1.11
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
pyfacer==0.0.4
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.12
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruff==0.11.8
safehttpx==0.1.6
safetensors==0.5.3
SAM-2==1.0
scikit-image==0.25.2
scikit-learn==1.6.1
scikit-video==1.1.11
scipy==1.15.2
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.27.0
setproctitle==1.3.6
setuptools==78.1.1
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
spaces==0.30.4
spandrel==0.4.0
stack-data==0.6.3
starlette==0.46.2
stringzilla==3.12.5
supervision==0.25.1
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.3.30
timm==1.0.9
tokenizers==0.20.1
tomlkit==0.12.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
triton==3.1.0
typer==0.15.3
typing_extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
validators==0.35.0
wandb==0.19.10
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
xformers==0.0.28.post3
zipp==3.21.0
